+++ tests/test_blurb.py
@@ -3,6 +3,14 @@ import pytest
 from blurb import blurb
 
 
+def test_section_names():
+    assert tuple(blurb.sections) == (
+        'Security', 'Core and Builtins', 'Library', 'Documentation',
+        'Tests', 'Build', 'Windows', 'macOS', 'IDLE', 'Tools/Demos',
+        'C API',
+    )
+
+
 UNCHANGED_SECTIONS = (
     "Library",
 )